0x00. C - Hello, World

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
   long int c;
   long long int d;
   float e;
-  printf("the size of a char: %lu byte(s)\n", sizeof(b));
-  printf("the size of an int: %lu byte(s)\n", sizeof(a));
-  printf("the size of a long int: %lu byte(s)\n", sizeof(c));
-  printf("the size of a long long int: %lu byte(s)\n", sizeof(d));
-  printf("the size of a float: %lu byte(s)\n", sizeof(e));
+  printf("the size of a char: %lu byte(s)\n", (unsigned long)sizeof(b));
+  printf("the size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+  printf("the size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+  printf("the size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+  printf("the size of a float: %lu byte(s)\n", (unsigned long)sizeof(e));
   return(0);
 }
